requirements.txt
transformers
torch
Pillow
sentence-transformers
scikit-learn
python-doctr[torch]
accelerate
bitsandbytes
python-dotenv
torchvision
